requirements.txt
fastapi
uvicorn
pandas
numpy
plotly
paho-mqtt
ray
scikit-learn
